requirements.txt
line-bot-sdk
boto3
datetime
awswrangler